2.0.0.120: Added ability to set row height to fit to chart height, or use a prescribed height for both column scenarios.
This introduces new properties and some re-wording of existing ones, to better clarify their behaviour. The user can now effectively mix-and match prescribed heights and widths, and managing the opposing direction of the grid to suit their needs.

There has been a fair bit of refactoring as I took this opportunity to move some more calculations from the ViewModelHandler into the SmallMultiplesHelper (with an eventualy view to having this fairly modular for other purposes).

There's a bug in here where the # of columns per row won't allow the placeholder value in the properties pane and now defaults to the max int value. This is a regression and needs fixing, but I want to get these changes into source control before I tackle it. #3 will stay open until I can sort it out.